Split up MultiIde.hs into many files (#19198)
* Split up MultiIde.hs into many files

* Further split SubIde
