+++ sdk/compiler/damlc/lib/DA/Cli/Damlc/Command/MultiIde/Prefixing.hs
@@ -3,15 +3,9 @@
 
 {-# LANGUAGE PolyKinds #-}
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE ApplicativeDo #-}
-{-# LANGUAGE RankNTypes #-}
-{-# LANGUAGE TypeFamilies #-}
-{-# LANGUAGE DisambiguateRecordFields #-}
-{-# LANGUAGE AllowAmbiguousTypes #-}
-{-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE GADTs #-}
 
--- This module handles prefixing of identifiers generated by the SubIDEs, so that they are unique to the client, but are returned in the same form to the servers.
+-- This module handles prefixing of identifiers generated by the SubIdes, so that they are unique to the client, but are returned in the same form to the servers.
 -- It has been implemented to be stateless for simplicity
 module DA.Cli.Damlc.Command.MultiIde.Prefixing (
   addProgressTokenPrefixToClientMessage,
@@ -30,38 +24,38 @@ import qualified Language.LSP.Types.Lens as LSP
 import DA.Cli.Damlc.Command.MultiIde.Types
 
 -- ProgressToken Prefixing
--- Progress tokens can be created on both client and subIDEs. They are then reported on by a subIDE, and cancelled by the client.
---   We need to avoid collisions between tokens created by different subIDEs.
--- Server created progress tokens use the SWindowWorkDoneProgressCreate notification. We prefix these tokens uniquely to the subIDE that created them.
+-- Progress tokens can be created on both client and subIdes. They are then reported on by a subIde, and cancelled by the client.
+--   We need to avoid collisions between tokens created by different subIdes.
+-- Server created progress tokens use the SWindowWorkDoneProgressCreate notification. We prefix these tokens uniquely to the subIde that created them.
 --   Handled in addProgressTokenPrefixToServerMessage
 -- Client created progress tokens are created alongside requests, `getProgressLenses` creates lenses for these tokens. We prefix these with `client`
 --   All prefixed tokens also include their original type, so the transformation can be safely reversed.
 --   Handled in addProgressTokenPrefixToClientMessage
 -- Progress messages
---   Sent from subIDEs, these are always forwarded to the client. However, depending on what created the Token, the prefix will need changing.
---   If the subIDE created the token, the progress message token will be unprefixed.
+--   Sent from subIdes, these are always forwarded to the client. However, depending on what created the Token, the prefix will need changing.
+--   If the subIde created the token, the progress message token will be unprefixed.
 --     We check the first character of the token. Neither server or client can create tokens starting with non-hex character. Our prefixing starts with `i` or `t`
---     If there is no prefix, we add the prefix of the sending subIDE. No subIDE should ever give progress reports for tokens created by other subIDEs.
+--     If there is no prefix, we add the prefix of the sending subIde. No subIde should ever give progress reports for tokens created by other subIdes.
 --   If the client created the token, it will have the client prefix. We detect this and remove it, so the client sees the exact name it created.
 --   Handled in addProgressTokenPrefixToServerMessage
 -- Cancel messages
---   Sent by the client, forwarded to either all or a specific subIDE.
---   If the token was created by the client, it will have no prefix. We add the prefix and broadcast to all subIDEs
---     This is safe because it is impossible for a subIDE to generate a token matching the prefixed client token, so the correct subIDE will delete the token, and the rest will ignore
---   If the token was created by a subIDE, it will have the subIDE's unique prefix. We strip this from the message, and return it to MultiIde.hs.
+--   Sent by the client, forwarded to either all or a specific subIde.
+--   If the token was created by the client, it will have no prefix. We add the prefix and broadcast to all subIdes
+--     This is safe because it is impossible for a subIde to generate a token matching the prefixed client token, so the correct subIde will delete the token, and the rest will ignore
+--   If the token was created by a subIde, it will have the subIde's unique prefix. We strip this from the message, and return it to MultiIde.hs.
 --     The message handling there will lookup the IDE matching this prefix and send the message to it. If no IDE exists, we can safely drop the message, as the IDE has been removed.
 
 -- | Convenience type for prefixes
 data ProgressTokenPrefix
-  = SubIDEPrefix T.Text
+  = SubIdePrefix T.Text
   | ClientPrefix
 
 progressTokenPrefixToText :: ProgressTokenPrefix -> T.Text
-progressTokenPrefixToText (SubIDEPrefix t) = t
+progressTokenPrefixToText (SubIdePrefix t) = t
 progressTokenPrefixToText ClientPrefix = "client"
 
 progressTokenPrefixFromMaybe :: Maybe T.Text -> ProgressTokenPrefix
-progressTokenPrefixFromMaybe = maybe ClientPrefix SubIDEPrefix
+progressTokenPrefixFromMaybe = maybe ClientPrefix SubIdePrefix
 
 -- | Reversible ProgressToken prefixing
 -- Given ProgressTokens can be int or text, we encode them as text as well as a tag to say if the original was an int
@@ -74,7 +68,7 @@ addProgressTokenPrefix prefix (LSP.ProgressTextToken t) = LSP.ProgressTextToken 
 progressTokenSplitPrefix :: T.Text -> (T.Text, Maybe T.Text)
 progressTokenSplitPrefix = bimap T.tail (mfilter (/="client") . Just) . swap . T.breakOn "-"
 
--- Removes prefix, returns the subIDE prefix if the token was created by a subIDE
+-- Removes prefix, returns the subIde prefix if the token was created by a subIde
 stripProgressTokenPrefix :: LSP.ProgressToken -> (LSP.ProgressToken, Maybe ProgressTokenPrefix)
 stripProgressTokenPrefix (LSP.ProgressTextToken (T.uncons -> Just ('i', rest))) =
   bimap (LSP.ProgressNumericToken . read . T.unpack) (Just . progressTokenPrefixFromMaybe) $ progressTokenSplitPrefix rest
@@ -82,19 +76,19 @@ stripProgressTokenPrefix (LSP.ProgressTextToken (T.uncons -> Just ('t', rest))) 
   bimap LSP.ProgressTextToken (Just . progressTokenPrefixFromMaybe) $ progressTokenSplitPrefix rest
 stripProgressTokenPrefix t = (t, Nothing)
 
--- Prefixes the SWindowWorkDoneProgressCreate and SProgress messages from subIDE. Rest are unchanged.
+-- Prefixes the SWindowWorkDoneProgressCreate and SProgress messages from subIde. Rest are unchanged.
 addProgressTokenPrefixToServerMessage :: T.Text -> LSP.FromServerMessage -> LSP.FromServerMessage
 addProgressTokenPrefixToServerMessage prefix (LSP.FromServerMess LSP.SWindowWorkDoneProgressCreate req) =
-  LSP.FromServerMess LSP.SWindowWorkDoneProgressCreate $ req & LSP.params . LSP.token %~ addProgressTokenPrefix (SubIDEPrefix prefix)
+  LSP.FromServerMess LSP.SWindowWorkDoneProgressCreate $ req & LSP.params . LSP.token %~ addProgressTokenPrefix (SubIdePrefix prefix)
 addProgressTokenPrefixToServerMessage prefix (LSP.FromServerMess LSP.SProgress notif) =
   case stripProgressTokenPrefix $ notif ^. LSP.params . LSP.token of
-    -- ProgressToken was created by this subIDE, add its usual prefix
+    -- ProgressToken was created by this subIde, add its usual prefix
     (unprefixedToken, Nothing) ->
-      let prefixedToken = addProgressTokenPrefix (SubIDEPrefix prefix) unprefixedToken
+      let prefixedToken = addProgressTokenPrefix (SubIdePrefix prefix) unprefixedToken
        in LSP.FromServerMess LSP.SProgress $ notif & LSP.params . LSP.token .~ prefixedToken
     -- ProgressToken was created by client, send back the unprefixed token
     (unprefixedToken, Just ClientPrefix) -> LSP.FromServerMess LSP.SProgress $ notif & LSP.params . LSP.token .~ unprefixedToken
-    (_, Just (SubIDEPrefix t)) -> error $ "SubIDE with prefix " <> T.unpack t <> " is somehow aware of its own prefixing. Something is very wrong."
+    (_, Just (SubIdePrefix t)) -> error $ "SubIde with prefix " <> T.unpack t <> " is somehow aware of its own prefixing. Something is very wrong."
 addProgressTokenPrefixToServerMessage _ msg = msg
 
 -- Prefixes client created progress tokens for all requests that can create them.
@@ -168,23 +162,23 @@ getProgressLenses = \case
       => ProgressLenses m
     both = ProgressLenses (Just $ Lens $ LSP.params . LSP.workDoneToken) (Just $ Lens $ LSP.params . LSP.partialResultToken)
 
--- strips and returns the subIDE prefix from cancel messages. Gives Nothing for client created tokens
+-- strips and returns the subIde prefix from cancel messages. Gives Nothing for client created tokens
 stripWorkDoneProgressCancelTokenPrefix
   :: LSP.NotificationMessage 'LSP.WindowWorkDoneProgressCancel
   -> (LSP.NotificationMessage 'LSP.WindowWorkDoneProgressCancel, Maybe T.Text)
 stripWorkDoneProgressCancelTokenPrefix notif =
   case stripProgressTokenPrefix $ notif ^. LSP.params . LSP.token of
-    -- Token was created by the client, add the client prefix and broadcast to all subIDEs
+    -- Token was created by the client, add the client prefix and broadcast to all subIdes
     (unprefixedToken, Nothing) ->
       let prefixedToken = addProgressTokenPrefix ClientPrefix unprefixedToken
        in (notif & LSP.params . LSP.token .~ prefixedToken, Nothing)
-    -- Created by subIDE, strip the prefix and send to the specific subIDE that created it.
-    (unprefixedToken, Just (SubIDEPrefix prefix)) -> (notif & LSP.params . LSP.token .~ unprefixedToken, Just prefix)
+    -- Created by subIde, strip the prefix and send to the specific subIde that created it.
+    (unprefixedToken, Just (SubIdePrefix prefix)) -> (notif & LSP.params . LSP.token .~ unprefixedToken, Just prefix)
     (_, Just ClientPrefix) -> error "Client attempted to cancel a ProgressToken with the client prefix, which it should not be aware of. Something went wrong."
 
 -- LspId Prefixing
 
--- We need to ensure all IDs from different subIDEs are unique to the client, so we prefix them.
+-- We need to ensure all IDs from different subIdes are unique to the client, so we prefix them.
 -- Given IDs can be int or text, we encode them as text as well as a tag to say if the original was an int
 -- Such that IdInt 10       -> IdString "iPREFIX-10"
 --       and IdString "hello" -> IdString "tPREFIX-hello"
@@ -206,7 +200,7 @@ stripLspPrefix (LSP.IdString (T.uncons -> Just ('t', rest))) = LSP.IdString $ T.
 stripLspPrefix t = t
 
 -- Prefixes applied to builtin and custom requests. Notifications do not have ids, responses do not need this logic.
-addLspPrefixToServerMessage :: SubIDEInstance -> LSP.FromServerMessage -> LSP.FromServerMessage
+addLspPrefixToServerMessage :: SubIdeInstance -> LSP.FromServerMessage -> LSP.FromServerMessage
 addLspPrefixToServerMessage _ res@(LSP.FromServerRsp _ _) = res
 addLspPrefixToServerMessage ide res@(LSP.FromServerMess method params) =
   case LSP.splitServerMethod method of
